requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
DateTime==4.9
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.103.2
frozenlist==1.4.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.6
ipython==8.18.1
jedi==0.19.1
json5==0.9.14
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.26.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
shimoku-api-python==1.4.1
shimoku-components-catalog==0.2.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
starlette-graphene3==0.6.0
strawberry-graphql==0.198.0
tenacity==8.1.0
tqdm==4.64.1
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
wcwidth==0.2.12
websockets==11.0.3
yarl==1.9.3
zope.interface==6.1
